+++ app/controllers/api/application_controller.rb
@@ -1,0 +1,5 @@
+require 'rapido'
+
+class Api::ApplicationController < ApplicationController
+  include Rapido::Auth::ApiKey
+end
