+++ app/controllers/api/steps_controller.rb
@@ -1,43 +1,11 @@
-class Api::StepsController < ApplicationController
-  before_action :load_workflow
+class Api::StepsController < Api::ApplicationController
+  include Rapido::Controller
 
-  rescue_from Workflow::UnmatchableDataError do |e|
-    render json: { error: "No matching step found.", facts: fact_params.to_h }, status: 422
-  end
+  owner_class :workflow
+  owner_lookup_param :workflow_token
+  owner_lookup_field :token
 
-  rescue_from Step::Parts::AnswerNotFound do |e|
-    render json: { error: e }, status: 500
-  end
+  resource_lookup_param :token
+  resource_permitted_params [:token, :text, :conditions, :callout, :callout_method, :callout_body]
 
-  rescue_from Step::Answers::AnswerNotFound do |e|
-    render json: { error: e }, status: 500
-  end
-
-  def create
-    Rails.logger.info "New step request for workflow #{params[:workflow_id]}: #{fact_params}"
-    step = @workflow.match(fact_params)
-    render json: step.to_h(fact_params.to_h)
-  end
-
-  private
-
-  def load_workflow
-    @workflow = Workflow.find_by(token: params[:workflow_id])
-    return if @workflow
-    render json: { error: "Workflow not found" }, status: :not_found
-  end
-
-  def fact_params
-    params[:facts].present? ? params.require(:facts).permit! : {}
-  end
-
-  def step
-    @step = {}
-    @step[:text] = "My first name is [:user_first_name], and my last name is [:user_last_name]."
-    @step[:answers] = {
-      user_first_name: { input_type: "short_text" },
-      user_last_name: { input_type: "short_text" }
-    }
-    @step
-  end
 end
