+++ app/models/account.rb
@@ -1,0 +1,7 @@
+class Account < ActiveRecord::Base
+  has_many :workflows
+
+  before_create do
+    self.api_key ||= SecureRandom::uuid()
+  end
+end
