+++ features/api/steps.feature
@@ -1,159 +1,98 @@
-Feature: Steps API
+Feature: Steps
 
   Background:
-    Given I am Ragnar Lothbrok
+    Given I am Andy Developer
     And I send and accept JSON
 
-  Scenario: Create inline step with multiple answers
-    When I send a POST request to "/api/workflows/abc/steps" with the following:
+  Scenario: System prevents access without api_key
+    When I send a POST request to "/api/workflows/okayflow/steps?api_key=12345" with the following:
     """
     {
-      "facts": {}
+      "token": "newstep",
+      "text": "okay text",
+      "conditions": "real="
     }
     """
-    Then the response status should be "200"
+    Then the response status should be "401"
     And the JSON response should be:
     """
     {
-      "token": "intro",
-      "text": "My first name is {{?user_first_name}}, and my last name is {{?user_last_name}}, and I am {{?user_age}} years old. Generally I prefer {{?user_style_preference}} wines.",
-      "parts": [
-        { "type": "text", "content": "My first name is " },
-        { "type": "short_text", "name": "user_first_name", "text_field_type": "text", "characters": 8 },
-        { "type": "text", "content": "," },
-        { "type": "text", "content": " and my last name is " },
-        { "type": "short_text", "name": "user_last_name", "text_field_type": "text", "characters": 10 },
-        { "type": "text", "content": "," },
-        { "type": "text", "content": " and I am " },
-        { "type": "short_text", "name": "user_age", "characters": 3, "text_field_type": "number" },
-        { "type": "text", "content": " years old." },
-        { "type": "text", "content": " Generally I prefer " },
-        {
-          "type": "select",
-          "name": "user_style_preference",
-          "options": [
-            { "text": "red", "value": "red" },
-            { "text": "white", "value": "white" }
-          ]
-        },
-        { "type": "text", "content": " wines." }
-      ]
+      "error": "Request denied."
     }
     """
 
-  Scenario: Create inline step with a single option answer and a user fact
-    When I send a POST request to "/api/workflows/abc/steps" with the following:
+  Scenario: Create a step
+    When I send a POST request to "/api/workflows/okayflow/steps?api_key=ABCDE" with the following:
     """
     {
-      "facts": {
-        "user_first_name": "Giles",
-        "user_last_name": "Rotherbrok",
-        "user_style_preference": "white"
-      }
+      "token": "newstep",
+      "text": "new text",
+      "conditions": "new="
     }
     """
-    Then the response status should be "200"
+    Then the response status should be "201"
     And the JSON response should be:
     """
     {
-      "token": "hello",
-      "text": "Hello {{@user_first_name}}, would you like any recommendations?\n{{?user_desires_recommendations}}",
-      "parts": [
-        { "type": "text", "content": "Hello " },
-        { "type": "text", "content": "Giles" },
-        { "type": "text", "content": "," },
-        { "type": "text", "content": " would you like any recommendations?" },
-        { "type": "newline" },
-        {
-          "type": "select",
-          "name": "user_desires_recommendations",
-          "options": [
-            { "text": "Yes", "value": "yes" },
-            { "text": "No", "value": "no" }
-          ]
-        }
-      ]
+      "token": "newstep",
+      "text": "new text",
+      "conditions": "new="
     }
     """
 
-  Scenario: Conditions can process a negative value
-    When I send a POST request to "/api/workflows/abc/steps" with the following:
-    """
-    {
-      "facts": {
-        "user_first_name": "Giles",
-        "user_last_name": "Rotherbrok",
-        "user_style_preference": "white",
-        "user_desires_recommendations": "no"
-      }
-    }
-    """
+  Scenario: Get a step
+    When I send a GET request to "/api/workflows/okayflow/steps/okaystep?api_key=ABCDE"
     Then the response status should be "200"
     And the JSON response should be:
     """
     {
-      "token": "finethen",
-      "text": "Well, fine then.",
-      "parts": [
-        { "type": "text", "content": "Well," },
-        { "type": "text", "content": " fine then." }
-      ]
+      "token": "okaystep",
+      "text": "okay text",
+      "conditions": "hmm="
     }
     """
 
-  Scenario: Create step with no answers and lots of breaking punctuation
-    When I send a POST request to "/api/workflows/abc/steps" with the following:
+  Scenario: Get all steps
+    When I send a GET request to "/api/workflows/okayflow/steps?api_key=ABCDE"
+    Then the response status should be "200"
+    And the JSON response should be:
+    """
+    [
+      {
+        "token": "okaystep",
+        "text": "okay text",
+        "conditions": "hmm="
+      }
+    ]
+    """
+
+  Scenario: Patch a step
+    When I send a PATCH request to "/api/workflows/okayflow/steps/okaystep?api_key=ABCDE"
     """
     {
-      "facts": {
-        "user_first_name": "Giles",
-        "user_last_name": "Rotherbrok",
-        "user_style_preference": "white",
-        "user_desires_recommendations": "yes"
-      }
+      "token": "updatedstep",
+      "text": "new text",
+      "conditions": "ahh="
     }
     """
     Then the response status should be "200"
     And the JSON response should be:
     """
     {
-      "token": "really",
-      "text": "Really? Splendid! Superb. I recommend the {{@recommendation}}.",
-      "parts": [
-        { "type": "text", "content": "Really?" },
-        { "type": "text", "content": " Splendid!" },
-        { "type": "text", "content": " Superb." },
-        { "type": "text", "content": " I recommend the " },
-        { "type": "text", "content": "Russian River Chardonnay" },
-        { "type": "text", "content": "." }
-      ]
+      "token": "updatedstep",
+      "text": "new text",
+      "conditions": "ahh="
     }
     """
 
-  Scenario: Create step with no answers and lots of breaking punctuation
-    When I send a POST request to "/api/workflows/abc/steps" with the following:
-    """
-    {
-      "facts": {
-        "user_first_name": "Giles",
-        "user_last_name": "Rotherbrok",
-        "user_style_preference": "red",
-        "user_desires_recommendations": "yes"
-      }
-    }
-    """
+  Scenario: Delete a step
+    When I send a DELETE request to "/api/workflows/okayflow/steps/okaystep?api_key=ABCDE"
     Then the response status should be "200"
     And the JSON response should be:
     """
     {
-      "token": "recommendation",
-      "text": "I recommend the {{@recommendation}}.\n{{?recommendation}}",
-      "parts": [
-        { "type": "text", "content": "I recommend the " },
-        { "type": "text", "content": "Pinot Noir" },
-        { "type": "text", "content": "." },
-        { "type": "newline" },
-        { "type": "hidden", "name": "recommendation", "value": "Pinot Noir" }
-      ]
+      "token": "okaystep",
+      "text": "okay text",
+      "conditions": "hmm="
     }
     """
