+++ features/personas/developers.rb
@@ -1,0 +1,11 @@
+Cucumber::Persona.define "Andy Developer" do
+  account = Account.create!(api_key: "ABCDE")
+  workflow = account.workflows.create!(token: "okayflow")
+  workflow.steps.create!(token: 'okaystep', text: "okay text", conditions: "hmm=")
+end
+
+Cucumber::Persona.define "Betsy Developer" do
+  account = Account.create!(api_key: "XYZ")
+  workflow = account.workflows.create!(token: "betterflow")
+  workflow.steps.create!(token: 'betterstep', text: "Hmm!", conditions: "nothing=")
+end
