+++ features/step_definitions/persona_steps.rb
@@ -1,3 +1,7 @@
-Given(/^I am (.*)$/) do |name|
+Given(/^I am ([a-zA-z]+ [a-zA-z]+)$/) do |name|
+  Cucumber::Persona.find(name).create
+end
+
+Given(/^([a-zA-z]+ [a-zA-z]+) exists$/) do |name|
   Cucumber::Persona.find(name).create
 end
